UBUNTU: [Config] Include vmd in storage-core-modules udeb
BugLink: http://bugs.launchpad.net/bugs/1703339

This is a secondary PCIe host bridge which may have NVMe devices
behind it. The module is needed in the installer to install to
these disks.

Note that while theoretically non-disk devices such as NICs could
be connected in a VMD domain, Intel says they do not support such
usage at this time. If supported in the future vmd should be
moved to the kernel-image udeb.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>